% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krig.R
\name{krig}
\alias{krig}
\title{Krige soil data following the methodology of the John et al. (2007).}
\usage{
krig(soil, var, params = NULL, gridsize = 20, plotdim = NULL,
  breaks = krig_breaks(2, 320, 30), use_ksline = TRUE, quiet = FALSE)
}
\arguments{
\item{soil}{The data frame with the points, coords specified in the
columns \code{gx}, \code{gy}.}

\item{var}{A character vector giving the name of each column in the soil
dataset #' containing soil data to krige.}

\item{params}{If you want to pass specified kriging parameters; see
\code{\link[=krig_auto_params]{krig_auto_params()}} for each parameter.}

\item{gridsize}{Points are kriged to the center points of a grid of
this size.}

\item{plotdim}{Numeric vector giving x and y dimensions of the plot. If
\code{NULL} (default) it will be guessed. Otherwise, it must be of length 2 with
the format \code{c(x, y)}.}

\item{breaks}{Breaks/intervals used to calculate the semivariogram.}

\item{use_ksline}{Use the \code{\link[geoR:ksline]{geoR::ksline()}} function? Use \code{TRUE} to calculate
a "best" semivariogram based on default parameters via
\code{geoR::variogram()}]. Use \code{FALSE} to base calculation on parameters passed
to \code{params}.}

\item{quiet}{Use \code{TRUE} to suppresses messages.}
}
\value{
A list with the following items:
\itemize{
\item \code{df}: Data frame of kriged values (column z) at each grid point (x, y).
\item \code{df.poly}: Data frame of the polynomial surface fitted to the raw data.
\item \code{lambda}: The "lambda" value used in the Box-Cox transform of the raw
data.
\item \code{vg}: A list giving the variogram parameters used for the kriging.
\item \code{vm}: Minimum loss value returned from \code{\link[geoR:variofit]{geoR::variofit()}}.
}
}
\description{
Use \code{krig()} to krige soil data following the methodology of the John et al.
(2007). Features:
\itemize{
\item Tries to guess \code{plotdim}.
\item Informs the guessed \code{plotdim} and the \code{gridsize} provided.
\item Allows to suppress messages.
\item Is vectorized over the argument \code{var}, to easily iterate over multiple soil
variables.
\item Has a method for \code{summary()}(see examples).
\item Has a method for \code{to_df()} (see examples).
}
}
\section{Deprecated}{

\code{krig()} evolves from \code{GetKrigedSoil()}.  \code{GetKrigedSoil()} has been
deprecated, although it remains in the package as an internal function.
}

\section{Breaks default}{

The default breaks argument is set to have more points where the exponential
curve rises the most and fewer at large distances. This means that the curve
fitting is not overly biased by points beyond the effective maximum range.
}

\examples{
library(fgeo.tool)

# Using automated parameters
summary(krig(soil_fake, var = "c"))

# Now using custom parameters (arbitrary but based on automated kriging params)
params <- list(
  model = "circular", range = 100, nugget = 1000, sill = 46000, kappa = 0.5
)

# Also using not one but multiple soil variables
vars <- c("c", "p")
custom <- krig(soil_fake, vars, params = params, quiet = TRUE)
summary(custom)

to_df(custom, name = "soil_var")

tail(to_df(custom, item = "df.poly"))
}
\seealso{
\code{\link[geoR:variofit]{geoR::variofit()}}, \code{\link[geoR:variog]{geoR::variog()}}, \code{\link[geoR:as.geodata]{geoR::as.geodata()}},
\code{\link[geoR:ksline]{geoR::ksline()}}, \code{\link[geoR:krige.conv]{geoR::krige.conv()}}, \code{geoR::krige.control()}.

\code{\link[fgeo.tool:to_df.krig_lst]{fgeo.tool::to_df.krig_lst()}}.
}
\author{
Graham Zemunik (grah.zem@gmail.com).
}
