% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krig_breaks.R
\name{krig_breaks}
\alias{krig_breaks}
\title{Sequence of n exponentially increasing values.}
\usage{
krig_breaks(first, last, n)
}
\arguments{
\item{first, last}{Numeric; first and last value of the list.}

\item{n}{Length of returned value.}
}
\value{
A numeric vector of length \code{n}.
}
\description{
Sequence of n exponentially increasing values.
}
\examples{
out <- krig_breaks(1, 10, 10)
plot(out)
}
\author{
Graham Zemunik (grah.zem@gmail.com).
}
