% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports-reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_tibble}
\alias{assert_is_installed}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fgeo.tool}{\code{\link[fgeo.tool]{assert_is_installed}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

