% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soil_fake}
\alias{soil_fake}
\title{Fake soil data with multiple soil variables.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30 rows and 5 columns.}
\usage{
soil_fake
}
\description{
Fake soil data with multiple soil variables.
}
\examples{
str(soil_fake)
}
\keyword{datasets}
