% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.krig_lst.R
\name{summary.krig_lst}
\alias{summary.krig_lst}
\title{Summary objects of class "krig_lst".}
\usage{
\method{summary}{krig_lst}(object, ...)
}
\arguments{
\item{object}{The result of \code{krig()}.}

\item{...}{Arguments passed on to \code{base::summary}
\describe{
  \item{object}{an object for which a summary is desired.}
}}
}
\value{
Prints a cleaner version of \code{str()} and returns its input invisibly.
}
\description{
This method helps you to visualize the output of \code{krig()} (which outputs
objects of class "krig_lst"). It is similar to \code{\link[utils:str]{utils::str()}} but a little
cleaner. version of \code{str()}.
}
\examples{
result <- krig(soil_fake, c("c", "p"), quiet = TRUE)
summary(result)
str(result)
}
\concept{methods for common generics}
