% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_df.R
\name{to_df.krig_lst}
\alias{to_df.krig_lst}
\title{Dataframe objects of class "krig_lst".}
\usage{
\method{to_df}{krig_lst}(.x, name = "var", item = "df", ...)
}
\arguments{
\item{.x}{The output of \code{fgeo.krig::krig()}.`}

\item{name}{Name for the column to hold soil variable-names.}

\item{item}{Character string; either "df" or "df.poly".}

\item{...}{Arguments passed on to \code{to_df}
\describe{
  \item{.x}{An fgeo object of supported class.}
}}
}
\value{
A dataframe.
}
\description{
This method creates a dataframe from the output of \code{fgeo.krig::krig()}
(which is a list of class "krig_lst").
}
\examples{
\dontrun{
if (!requireNamespace("fgeo.krig")) {
  stop("This example requires fgeo.krig. Please install it")
}
library(fgeo.krig)

vars <- c("c", "p")
krig <- krig(soil_fake, vars, quiet = TRUE)
to_df(krig)
}
}
\seealso{
\code{\link[=to_df]{to_df()}}.
}
\concept{methods for fgeo generics}
